FIX: Added permission check to update_resource_type() [q10386]
git-svn-id: http://svn.resourcespace.com/svn/rs/trunk@11817 c08608d7-6e46-0410-86ca-f2a6f1370df5

+++ include/resource_functions.php
@@ -2518,21 +2518,28 @@ function get_themes_by_resource($ref)
 
 function update_resource_type($ref,$type)
 	{
+    if (checkperm("XU" . $type))
+        {
+        return false;
+        }
+        
 	sql_query("update resource set resource_type='$type' where ref='$ref'");
 	
 	# Clear data that is no longer needed (data/keywords set for other types).
 	sql_query("delete from resource_data where resource='$ref' and resource_type_field not in (select ref from resource_type_field where resource_type='$type' or resource_type=999 or resource_type=0)");
 	sql_query("delete from resource_keyword where resource='$ref' and resource_type_field>0 and resource_type_field not in (select ref from resource_type_field where resource_type='$type' or resource_type=999 or resource_type=0)");
 	sql_query("delete from resource_node where resource='$ref' and node>0 and node not in (select n.ref from node n left join resource_type_field rf on n.resource_type_field=rf.ref where rf.resource_type='$type' or rf.resource_type=999 or resource_type=0)");	
-        # Also index the resource type name, unless disabled
-        global $index_resource_type;
-        if ($index_resource_type)
-                {
-                $restypename=sql_value("select name value from resource_type where ref='" . escape_check($type) . "'","");
-                remove_all_keyword_mappings_for_field($ref,-2);
-                add_keyword_mappings($ref,$restypename,-2);
-                }
-        	
+    
+    # Also index the resource type name, unless disabled
+    global $index_resource_type;
+    if ($index_resource_type)
+            {
+            $restypename=sql_value("select name value from resource_type where ref='" . escape_check($type) . "'","");
+            remove_all_keyword_mappings_for_field($ref,-2);
+            add_keyword_mappings($ref,$restypename,-2);
+            }
+                
+    return true;    	
 	}
 	
 function relate_to_array($ref,$array)
